requirements.txt
flask>=3.0,<4.0
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
python-dotenv>=1.0,<2.0

selenium>=4.20,<5.0
playwright>=1.46,<2.0

groq>=0.11,<1.0
openai>=1.40,<2.0
google-generativeai>=0.7,<1.0

pydantic>=2.7,<3.0

